requirements.txt
absl-py==1.4.0
adal==1.2.7
altair==4.2.2
appdirs==1.4.4
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attr==0.3.1
attrs==23.1.0
azure-core==1.26.4
azure-storage-blob==12.16.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==5.0.1
blinker==1.6.2
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
expiringdict==1.2.2
Faker==18.9.0
favicon==0.7.0
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.39.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.10.0
google-api-python-client==2.87.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
graphviz==0.20.1
grpc-google-iam-v1==0.12.3
grpcio==1.54.2
grpcio-status==1.48.2
h5py==3.8.0
h5pyd==0.14.1
htbuilder==0.6.1
htmlmin==0.1.12
httplib2==0.22.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
ipython==8.14.0
ipyvizzu==0.15.0
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
keras==2.12.0
kiwisolver==1.4.4
label-studio==1.7.3
label-studio-converter==0.0.51
label-studio-tools==0.0.2
launchdarkly-server-sdk==7.5.0
libclang==16.0.0
lockfile==0.12.2
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
more-itertools==9.1.0
mpmath==1.3.0
msrest==0.7.1
msrestazure==0.6.4
multimethod==1.9.1
networkx==3.1
nltk==3.6.7
numpy==1.21.6
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==23.1
pandas==1.3.5
pandas-profiling==3.6.6
parso==0.8.3
pascal==0.0.1.dev1
pascal-voc-writer==0.1.4
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.3.0
Plost==0.2.5
plotly==5.14.1
pluggy==1.0.0
prettytable==3.7.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psycopg2-binary==2.9.6
ptyprocess==0.7.0
PuLP==2.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
pyecharts==2.0.3
Pygments==2.15.1
PyJWT==2.7.0
pymdown-extensions==10.0.1
Pympler==1.0.1
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.1.0
python-coveralls==2.9.3
python-dateutil==2.8.1
python-json-logger==2.0.4
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
requests==2.28.0
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
rich==13.3.5
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semver==2.13.0
sentry-sdk==1.25.0
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
soupsieve==2.4.1
sqlparse==0.4.4
st-annotated-text==4.0.0
st-vizzu==0.1
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.23.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-drawable-canvas==0.9.2
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-annotation==0.3.1
streamlit-image-coordinates==0.1.3
streamlit-img-label==0.1.1
streamlit-keyup==0.2.0
streamlit-labelstudio==0.0.3
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.6
streamlit-pandas-profiling==0.1.3
streamlit-tensorboard==0.0.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ua-parser==0.16.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
validators==0.20.0
visions==0.7.5
voc==0.1.6
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.4
wrapt==1.14.1
xmljson==0.2.0
ydata-profiling==4.1.2
zipp==3.15.0
